Renamed imports in tests, fixed syntax errors
